Moved scraper to helper file
